Changed Enemy movement a bit
I changed the enemy movement to, instead of only attacking when the player is close enough, they will now wait for x time after they spawn and then attack the player no matter the distance. This ensures that the player cannot just 'hide' on one side of the map and wait to heal before re-entering the battlefield.